==> python_NorthAmerica/3_candidates_creation.ipynb <==
import pandas as pd
import parameters_cand_NA as parameters
#from parameters import parameters
#from parameters_cand_NA import parameters as parameters
#from create_coregistration_input_data_example import *
import pygplates
import glob
import Utils_cand_NA as Utils

deposit_candidates = Utils.get_deposit_candidates()
deposit_candidates.to_csv('case_AREPS/coreg_input/NA/deposit_candidates.csv', index_label = 'index')

# import pandas as pd
# import parameters_cand_SA as parameters
# #from parameters import parameters
# #from create_coregistration_input_data_example import *
# import pygplates
# import glob
# import Utils_cand_SA as Utils

# deposit_candidates = Utils.get_deposit_candidates()
# deposit_candidates.to_csv('case_AREPS/coreg_input/SA/deposit_candidates.csv', index_label = 'index')
